+++ wxc_sdk/__init__.py
@@ -34,6 +34,7 @@ from .telephony import TelephonyApi
 from .tokens import Tokens
 from .webhook import WebhookApi
 from .workspace_locations import WorkspaceLocationApi
+from .workspace_personalization import WorkspacePersonalizationApi
 from .workspaces import WorkspacesApi
 from .workspace_settings import WorkspaceSettingsApi
 from dataclasses import dataclass
@@ -108,6 +109,8 @@ class WebexSimpleApi:
     workspaces: WorkspacesApi
     #: Workspace locations API; :class:`workspace_locations.WorkspaceLocationApi`
     workspace_locations: WorkspaceLocationApi
+    #: Workspace personalization API :class:workspace_personalization.WorkspacePersonalizationApi`
+    workspace_personalization: WorkspacePersonalizationApi
     #: Workspace setting API :class:`workspace_settings.WorkspaceSettingsApi`
     workspace_settings: WorkspaceSettingsApi
     #: :class:`rest.RestSession` used for all API requests
@@ -161,6 +164,7 @@ class WebexSimpleApi:
         self.webhook = WebhookApi(session=session)
         self.workspaces = WorkspacesApi(session=session)
         self.workspace_locations = WorkspaceLocationApi(session=session)
+        self.workspace_personalization = WorkspacePersonalizationApi(session=session)
         self.workspace_settings = WorkspaceSettingsApi(session=session)
         self.session = session